handleIsError In tablayout
