+++ src/app/components/Tabs/TabLayout.tsx
@@ -5,10 +5,11 @@ import TabPanel from '@app/components/Tabs/TabPanel';
 import { TabProps as TabsPropsMui } from '@mui/material';
 
 interface optionsTabLayout extends TabsPropsMui {
-  label: string;
   renderelement: JSX.Element;
   isError?: boolean;
-  rest?: TabsPropsMui;
+  formFields?: Array<string>;
+  //TODO: tipar correctamente formErrors y no de modo generico
+  formErrors?: AnyValue
 }
 
 type TabPros = {
@@ -21,12 +22,33 @@ const TabLayout = ({ options }: TabPros) => {
     setTabValue(newValue);
   };
 
+  /** Devuelve boolean para TabLayout. Revisa si hay un form error dentro de los formsFields
+  * @param formFields
+  * @param formErrors
+  */
+  const handleIsError = (formFields: AnyValue, formErrors: AnyValue) => {
+
+    if (formErrors === null || formErrors === undefined) return false
+
+    const keysOfErrors = Object.keys(formErrors);
+    const auxilarArray = [];
+
+    for (const fieldOfForm of formFields) {
+      const hasError = keysOfErrors.includes(fieldOfForm);
+      auxilarArray.push(hasError);
+    }
+    const result = auxilarArray.filter(el => el == true);
+    return result.length > 0 ? true : false;
+  };
+
   return (
     <>
       <Tabs value={tabValue} onChange={handleChangeTab} scrollButtons allowScrollButtonsMobile variant='scrollable'>
         {options.map((el: optionsTabLayout, index: number) => {
-          const { ...rest } = el;
-          return <Tab key={index} value={index} {...rest} isError={el?.isError} />;
+          const { formFields, formErrors, ...rest } = el;
+          const isError = handleIsError(formFields, formErrors)
+
+          return <Tab key={index} value={index} {...rest} isError={isError} />;
         })}
       </Tabs>
       {options.map((el: optionsTabLayout, index: number) => {
